Update config.py
line49: added a directory for the config file. It will live inside the same folder as the module package (/opt/gls/clarity/customextensions/genologics/genologics.conf). We do not have write-access to /etc and our glsapiuser does not have a home directory. I am unsure where ~/.genologicsrc would be when running the script using the automation worker. I believe there is an automation worker user named glsai, but it is hidden for us. So this is a good workaround location for the config file, I think.

This change has been implemented on test-code on the test server. Confirmed working (ie no configfile errors).
@ajkarloss @georgemarselis-nvi